requirements.txt
Flask==2.0.3
python-dotenv==0.12.0
Flask-Bootstrap4==4.0.2
flask-sqlalchemy==2.5.1
flask-migrate==2.5.2
waitress==1.4.4
Jinja2==3.0.3
Werkzeug==2.0.3
MarkupSafe==2.0.1
SQLAlchemy==1.4.46
